Rewrite AdService to use a Facade pattern
The problem with the previous `AdRequestService`  implementation was that it was doing too much, it fetched either remote content or falled back to local., in case it failed.

With this implementation the `AdService` is only responsible for fetching remote content, and `AdPersistenceService` is responsible for fetching local data.  The `AdsFacade` is an abstraction around the two services.